fix win build

+++ src/cupoch/integration/scalable_tsdfvolume.cu
@@ -161,7 +161,7 @@ struct extract_pointcloud_functor {
                 (p_locate - index0.cast<float>() * volume_unit_length_) /
                 voxel_length_;
         for (int i = 0; i < 3; i++) {
-            idx0(i) = (int)floor(p_grid(i));
+            idx0(i) = (int)floorf(p_grid(i));
             if (idx0(i) < 0) idx0(i) = 0;
             if (idx0(i) >= resolution_)
                 idx0(i) = resolution_ - 1;
